requirements.txt
alembic==1.4.2
certifi==2020.4.5.2
click==7.1.2
cloudinary==1.21.0
Flask==1.1.2
Flask-Cors==3.0.8
flask-marshmallow==0.12.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
mysql-client==0.0.1
Pillow==7.1.2
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==1.0.1
